# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Personas",
        "version": "1.0"
    },
    "paths": {
        "/api/persona": {
            "get": {
                "tags": [
                    "Persona"
                ],
                "summary": "Mostrar personas",
                "operationId": "fa18ac5b05ac1b0c1a4b9992b86d4a59",
                "responses": {
                    "200": {
                        "description": "Mostrar todas las personas."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            },
            "post": {
                "tags": [
                    "Persona"
                ],
                "summary": "Lista de personas",
                "description": "Lista de personas",
                "operationId": "postPersona",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Noe"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "descripcion de la persona"
                                    },
                                    "user_id": {
                                        "type": "string",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/api/persona/{id}": {
            "get": {
                "tags": [
                    "Persona"
                ],
                "summary": "Optener persona por id",
                "description": "Optener persona por id",
                "operationId": "getPersona",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id,\n     *      eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Persona"
                ],
                "summary": "Lista de personas",
                "description": "Lista de personas",
                "operationId": "updatePersona",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id,\n     *      eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juliaca"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Juliaca"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Persona"
                ],
                "summary": "Lista de personas",
                "description": "Lista de personas",
                "operationId": "deletePersona",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id,\n     *      eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        }
    }
}